Solve DailyKw issue
At the start of the simulation GLM assigns the Kw value to all layers (glm_init.c, l. 644):
for (i = ; i < MaxLayers; i++) Lake[i].ExtcCoefSW = Kw;
However, when using a Kw_file, at the start of each day the value of DailyKw is assigned to cells from bottom to surface (not all layers) (glm_model.c, l. 559 and others):
for (i = botmLayer; i <= surfLayer; i ++) Lake[i].ExtcCoefSW = DailyKw;

As a result, when the surface layer splits, the new layers do not use the daily Kw value, but the Kw value specified at the start of the simulation or the latest ExtcCoefSW value used by that cell.

The proposed change solves the issue.